Media Endpoint improvements
For the code, media related stuff is now in its own controller

Added support for querying the most recent file uploaded
